test: expect HTTP 404 on missing gateway resources
When request to a PublicGateway fails due to the lack
of valid muxer or of current path prefix is not on the Paths whitelist
for the hostname we return a vanilla HTTP 404 Not Found error.

This ensures that is the case, and also adds tests to safeguard
against regressions in the future.

License: MIT
Signed-off-by: Marcin Rataj <lidel@lidel.org>